Revert "[WIP] cmd/geth: add status command (#251)"
Reason: I accidentally left '[WIP]' in the subject...

This reverts commit 9ae70d9a9159bef2ebc0c6e76b0e885c8653acef.
