+++ cmd/geth/flag.go
@@ -95,7 +95,7 @@ var reservedChainIDS = map[string]bool{
 }
 
 var testnetChaidIDs = map[string]bool{
-	"morden":  true,
+	"morden": true,
 	"testnet": true,
 }
 
@@ -307,13 +307,13 @@ func MakeAccountManager(ctx *cli.Context) *accounts.Manager {
 		scryptP = accounts.LightScryptP
 	}
 
-	// in case --chain-config is set, we need to read file to store chain subdir as global
 	if ctx.GlobalIsSet(aliasableName(UseChainConfigFlag.Name, ctx)) {
-		mustMakeSufficientChainConfig(ctx)
+		mustMakeSufficientConfiguration(ctx) // parses flags and set global chain ID
 	}
 	datadir := MustMakeChainDataDir(ctx)
 
 	keydir := filepath.Join(datadir, "keystore")
+
 	if path := ctx.GlobalString(aliasableName(KeyStoreDirFlag.Name, ctx)); path != "" {
 		af, e := filepath.Abs(path)
 		if e != nil {
@@ -383,6 +383,223 @@ func MakePasswordList(ctx *cli.Context) []string {
 	return lines
 }
 
+// migrateExistingDirToClassicNamingScheme renames default base data directory ".../Ethereum" to ".../EthereumClassic", pending os customs, etc... ;-)
+///
+// Check for preexisting **Un-classic** data directory, ie "/home/path/to/Ethereum".
+// If it exists, check if the data therein belongs to Classic blockchain (ie not configged as "ETF"),
+// and rename it to fit Classic naming convention ("/home/path/to/EthereumClassic") if that dir doesn't already exist.
+// This case only applies to Default, ie when a user **doesn't** provide a custom --datadir flag;
+// a user should be able to override a specified data dir if they want.
+func migrateExistingDirToClassicNamingScheme(ctx *cli.Context) error {
+
+	ethDataDirPath := common.DefaultUnclassicDataDir()
+	etcDataDirPath := common.DefaultDataDir()
+
+	// only if default <EthereumClassic>/ datadir doesn't already exist
+	if _, err := os.Stat(etcDataDirPath); err == nil {
+		// classic data dir already exists
+		glog.V(logger.Debug).Infof("Using existing ETClassic data directory at: %v\n", etcDataDirPath)
+		return nil
+	}
+
+	ethChainDBPath := filepath.Join(ethDataDirPath, "chaindata")
+	if isTestMode(ctx) {
+		ethChainDBPath = filepath.Join(ethDataDirPath, "testnet", "chaindata")
+	}
+
+	// only if ETHdatadir chaindb path DOES already exist, so return nil if it doesn't;
+	// otherwise NewLDBDatabase will create an empty one there.
+	// note that this uses the 'old' non-subdirectory way of holding default data.
+	// it must be called before migrating to subdirectories
+	// NOTE: Since ETH stores chaindata by default in Ethereum/geth/..., this path
+	// will not exist if the existing data belongs to ETH, so it works as a valid check for us as well.
+	if _, err := os.Stat(ethChainDBPath); os.IsNotExist(err) {
+		glog.V(logger.Debug).Infof(`No existing default chaindata dir found at: %v
+		  	Using default data directory at: %v`,
+			ethChainDBPath, etcDataDirPath)
+		return nil
+	}
+
+	foundCorrectLookingFiles := []string{}
+	requiredFiles := []string{"LOG", "LOCK", "CURRENT"}
+	for _, f := range requiredFiles {
+		p := filepath.Join(ethChainDBPath, f)
+		if _, err := os.Stat(p); os.IsNotExist(err) {
+			glog.V(logger.Debug).Infof(`No existing default file found at: %v
+		  	Using default data directory at: %v`,
+				p, etcDataDirPath)
+		} else {
+			foundCorrectLookingFiles = append(foundCorrectLookingFiles, f)
+		}
+	}
+	hasRequiredFiles := len(requiredFiles) == len(foundCorrectLookingFiles)
+	if !hasRequiredFiles {
+		return nil
+	}
+
+
+	// check if there is existing etf blockchain data in unclassic default dir (ie /<home>/Ethereum)
+	chainDB, err := ethdb.NewLDBDatabase(ethChainDBPath, 0, 0)
+	if err != nil {
+		glog.V(logger.Error).Info(`Failed to check blockchain compatibility for existing Ethereum chaindata database at: %v
+		 	Using default data directory at: %v`,
+			err, etcDataDirPath)
+		return nil
+	}
+
+	defer chainDB.Close()
+
+
+	// Only move if defaulty ETC (mainnet or testnet).
+	// Get head block if testnet, fork block if mainnet.
+	hh := core.GetHeadBlockHash(chainDB) // get last block in fork
+	if ctx.GlobalBool(aliasableName(FastSyncFlag.Name, ctx)) {
+		hh = core.GetHeadFastBlockHash(chainDB)
+	}
+	if hh.IsEmpty() {
+		glog.V(logger.Debug).Info("There was no head block for the old database. It could be very young.")
+	}
+
+	hasRequiredForkIfSufficientHeight := true
+	if !hh.IsEmpty() {
+		// if head block < 1920000, then its compatible
+		// if head block >= 1920000, then it must have a hash matching required hash
+
+		// Use default configuration to check if known fork, if block 1920000 exists.
+		// If block1920000 doesn't exist, given above checks for directory structure expectations,
+		// I think it's safe to assume that the chaindata directory is just too 'young', where it hasn't
+		// synced until block 1920000, and therefore can be migrated.
+		conf := core.DefaultConfig
+		if isTestMode(ctx) {
+			conf = core.TestConfig
+		}
+
+		hf := conf.ForkByName("The DAO Hard Fork")
+		if hf == nil || hf.Block == nil || new(big.Int).Cmp(hf.Block) == 0 || hf.RequiredHash.IsEmpty(){
+			glog.V(logger.Debug).Info("DAO Hard Fork required hash not configured for database chain. Not migrating.")
+			return nil
+		}
+
+		b := core.GetBlock(chainDB, hh)
+		if b == nil {
+			glog.V(logger.Debug).Info("There was a problem checking the head block of old-namespaced database. The head hash was: %v", hh.Hex())
+			return nil
+		}
+
+		// if head block >= 1920000
+		if b.Number().Cmp(hf.Block) >= 0 {
+			// now, since we know that the height is bigger than the hardfork, we have to check that the db contains the required hardfork hash
+			glog.V(logger.Debug).Infof("Existing head block in old data dir has sufficient height: %v", b.String())
+
+			hasRequiredForkIfSufficientHeight = false
+			bf := core.GetBlock(chainDB, hf.RequiredHash)
+			// does not have required block by hash
+			if bf != nil {
+				glog.V(logger.Debug).Infof("Head block has sufficient height AND required hash: %v", b.String())
+				hasRequiredForkIfSufficientHeight = true
+			} else {
+				glog.V(logger.Debug).Infof("Head block has sufficient height but not required hash: %v", b.String())
+			}
+		// head block < 1920000
+		} else {
+			glog.V(logger.Debug).Infof("Existing head block in old data dir has INSUFFICIENT height to differentiate ETC/ETF: %v", b.String())
+		}
+	}
+
+	if hasRequiredForkIfSufficientHeight {
+		// if any of the LOG, LOCK, or CURRENT files are missing from old chaindata/, don't migrate
+		glog.V(logger.Info).Infof(`Found existing data directory named 'Ethereum' with default ETC chaindata.
+		  	Moving it from: %v, to: %v
+		  	To specify a different data directory use the '--datadir' flag.`,
+			ethDataDirPath, etcDataDirPath)
+		return os.Rename(ethDataDirPath, etcDataDirPath)
+	}
+
+	glog.V(logger.Debug).Infof(`Existing default Ethereum database at: %v isn't an Ethereum Classic default blockchain.
+	  	Will not migrate.
+	  	Using ETC chaindata database at: %v`,
+		ethDataDirPath, etcDataDirPath)
+	return nil
+}
+
+// migrateToChainSubdirIfNecessary migrates ".../EthereumClassic/nodes|chaindata|...|nodekey" --> ".../EthereumClassic/mainnet/nodes|chaindata|...|nodekey"
+func migrateToChainSubdirIfNecessary(ctx *cli.Context) error {
+	name := getChainConfigIDFromContext(ctx) // "mainnet", "morden", "custom"
+
+	datapath := mustMakeDataDir(ctx) // ".../EthereumClassic/ | --datadir"
+
+	subdirPath := MustMakeChainDataDir(ctx) // ie, <EthereumClassic>/mainnet
+
+	// check if default subdir "mainnet" exits
+	// NOTE: this assumes that if the migration has been run once, the "mainnet" dir will exist and will have necessary datum inside it
+	subdirPathInfo, err := os.Stat(subdirPath)
+	if err == nil {
+		// dir already exists
+		return nil
+	}
+	if subdirPathInfo != nil && !subdirPathInfo.IsDir() {
+		return fmt.Errorf(`%v: found file named '%v' in EthereumClassic datadir,
+			which conflicts with default chain directory naming convention: %v`, ErrDirectoryStructure, name, subdirPath)
+	}
+
+	// 3.3 testnet uses subdir '/testnet'
+	if testnetChaidIDs[name] {
+		exTestDir := filepath.Join(subdirPath, "../testnet")
+		exTestDirInfo, e := os.Stat(exTestDir)
+		if e != nil && os.IsNotExist(e) {
+			return nil // ex testnet dir doesn't exist
+		}
+		if !exTestDirInfo.IsDir() {
+			return nil // don't interfere with user *file* that won't be relevant for geth
+		}
+		return os.Rename(exTestDir, subdirPath) // /testnet -> /morden
+	}
+
+	// mkdir -p ".../mainnet"
+	if err := os.MkdirAll(subdirPath, 0755); err != nil {
+		return err
+	}
+
+	// move if existing (nodekey, dapp/, keystore/, chaindata/, nodes/) into new subdirectories
+	for _, dir := range []string{"dapp", "keystore", "chaindata", "nodes"} {
+
+		dirPath := filepath.Join(datapath, dir)
+
+		dirInfo, e := os.Stat(dirPath)
+		if e != nil && os.IsNotExist(e) {
+			continue // dir doesn't exist
+		}
+		if !dirInfo.IsDir() {
+			continue // don't interfere with user *file* that won't be relevant for geth
+		}
+
+		dirPathUnderSubdir := filepath.Join(subdirPath, dir)
+		if err := os.Rename(dirPath, dirPathUnderSubdir); err != nil {
+			return err
+		}
+	}
+
+	// ensure nodekey exists and is file (loop lets us stay consistent in form here, an keep options open for easy other files to include)
+	for _, file := range []string{"nodekey", "geth.ipc"} {
+		filePath := filepath.Join(datapath, file)
+
+		// ensure exists and is a file
+		fileInfo, e := os.Stat(filePath)
+		if e != nil && os.IsNotExist(e) {
+			continue
+		}
+		if fileInfo.IsDir() {
+			continue // don't interfere with user dirs that won't be relevant for geth
+		}
+
+		filePathUnderSubdir := filepath.Join(subdirPath, file)
+		if err := os.Rename(filePath, filePathUnderSubdir); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 // makeName makes the node name, which can be (in part) customized by the IdentityFlag
 func makeNodeName(version string, ctx *cli.Context) string {
 	name := fmt.Sprintf("Geth/%s/%s/%s", version, runtime.GOOS, runtime.Version())
@@ -406,23 +623,11 @@ func chainIdIsCustom(ctx *cli.Context) bool {
 	return false
 }
 
-// shouldAttemptDirMigration decides based on flags if
-// should attempt to migration from old (<=3.3) directory schema to new.
-func shouldAttemptDirMigration(ctx *cli.Context) bool {
-	if !ctx.GlobalIsSet(aliasableName(DataDirFlag.Name, ctx)) {
-		if !chainIdIsCustom(ctx) {
-			if !ctx.GlobalIsSet(aliasableName(UseChainConfigFlag.Name, ctx)) {
-				return true
-			}
-
-		}
-	}
-	return false
-}
-
 // MakeSystemNode sets up a local node, configures the services to launch and
 // assembles the P2P protocol stack.
 func MakeSystemNode(version string, ctx *cli.Context) *node.Node {
+
+	name := makeNodeName(version, ctx)
 
 	// global settings
 
@@ -434,67 +639,46 @@ func MakeSystemNode(version string, ctx *cli.Context) *node.Node {
 		miner.HeaderExtra = []byte(s)
 	}
 
-	// Data migrations if should.
-	if shouldAttemptDirMigration(ctx) {
-		// Rename existing default datadir <home>/<Ethereum>/ to <home>/<EthereumClassic>.
-		// Only do this if --datadir flag is not specified AND <home>/<EthereumClassic> does NOT already exist (only migrate once and only for defaulty).
-		// If it finds an 'Ethereum' directory, it will check if it contains default ETC or ETHF chain data.
-		// If it contains ETC data, it will rename the dir. If ETHF data, if will do nothing.
-		if migrationError := migrateExistingDirToClassicNamingScheme(ctx); migrationError != nil {
-			glog.Fatalf("%v: failed to migrate existing Classic database: %v", ErrDirectoryStructure, migrationError)
-		}
+	// Data migrations...
 
-		// Move existing mainnet data to pertinent chain-named subdir scheme (ie ethereum-classic/mainnet).
-		// This should only happen if the given (newly defined in this protocol) subdir doesn't exist,
-		// and the dirs&files (nodekey, dapp, keystore, chaindata, nodes) do exist,
-		if subdirMigrateErr := migrateToChainSubdirIfNecessary(ctx); subdirMigrateErr != nil {
-			glog.Fatalf("%v: failed to migrate existing data to chain-specific subdir: %v", ErrDirectoryStructure, subdirMigrateErr)
+	// Rename existing default datadir <home>/<Ethereum>/ to <home>/<EthereumClassic>.
+	// Only do this if --datadir flag is not specified AND <home>/<EthereumClassic> does NOT already exist (only migrate once and only for defaulty).
+	// If it finds an 'Ethereum' directory, it will check if it contains default ETC or ETHF chain data.
+	// If it contains ETC data, it will rename the dir. If ETHF data, if will do nothing.
+	if !ctx.GlobalIsSet(aliasableName(DataDirFlag.Name, ctx)) {
+		if !chainIdIsCustom(ctx) {
+			if !ctx.GlobalIsSet(aliasableName(UseChainConfigFlag.Name, ctx)) {
+				if migrationError := migrateExistingDirToClassicNamingScheme(ctx); migrationError != nil {
+					glog.Fatalf("%v: failed to migrate existing Classic database: %v", ErrDirectoryStructure, migrationError)
+				}
+			}
 		}
 	}
+	// Move existing mainnet data to pertinent chain-named subdir scheme (ie ethereum-classic/mainnet).
+	// This should only happen if the given (newly defined in this protocol) subdir doesn't exist,
+	// and the dirs&files (nodekey, dapp, keystore, chaindata, nodes) do exist,
+	if !ctx.GlobalIsSet(aliasableName(DataDirFlag.Name, ctx)) {
+		if !chainIdIsCustom(ctx) {
+			if !ctx.GlobalIsSet(aliasableName(UseChainConfigFlag.Name, ctx)) {
+				if subdirMigrateErr := migrateToChainSubdirIfNecessary(ctx); subdirMigrateErr != nil {
+					glog.Fatalf("%v: failed to migrate existing data to chain-specific subdir: %v", ErrDirectoryStructure, subdirMigrateErr)
+				}
+			}
+		}
+	}
+
+	// Makes sufficient configuration from JSON file or DB pending flags.
+	// Delegates flag usage.
+	config := mustMakeSufficientConfiguration(ctx)
+	logChainConfiguration(ctx, config)
 
 	// Avoid conflicting flags
 	if ctx.GlobalBool(DevModeFlag.Name) && isTestMode(ctx) {
 		glog.Fatalf("%v: flags --%v and --%v/--%v=morden are mutually exclusive", ErrInvalidFlag, DevModeFlag.Name, TestNetFlag.Name, ChainIDFlag.Name)
 	}
 
-	// Makes sufficient configuration from JSON file or DB pending flags.
-	// Delegates flag usage.
-	config := mustMakeSufficientChainConfig(ctx)
-	logChainConfiguration(ctx, config)
-
-	// Configure the Ethereum service
-	ethConf := mustMakeEthConf(ctx, config)
-
-	// Configure node's service container.
-	name := makeNodeName(version, ctx)
-	stackConf, shhEnable := mustMakeStackConf(ctx, name, config, ethConf)
-
-	// Assemble and return the protocol stack
-	stack, err := node.New(stackConf)
-	if err != nil {
-		glog.Fatalf("%v: failed to create the protocol stack: ", ErrStackFail, err)
-	}
-	if err := stack.Register(func(ctx *node.ServiceContext) (node.Service, error) {
-		return eth.New(ctx, ethConf)
-	}); err != nil {
-		glog.Fatalf("%v: failed to register the Ethereum service: ", ErrStackFail, err)
-	}
-	if shhEnable {
-		if err := stack.Register(func(*node.ServiceContext) (node.Service, error) { return whisper.New(), nil }); err != nil {
-			glog.Fatalf("%v: failed to register the Whisper service: ", ErrStackFail, err)
-		}
-	}
-
-	if ctx.GlobalBool(Unused1.Name) {
-		glog.V(logger.Info).Infoln(fmt.Sprintf("Geth started with --%s flag, which is unused by Geth Classic and can be omitted", Unused1.Name))
-	}
-
-	return stack
-}
-
-func mustMakeStackConf(ctx *cli.Context, name string, config *core.SufficientChainConfig, ethConf *eth.Config) (stackConf *node.Config, shhEnable bool) {
 	// Configure the node's service container
-	stackConf = &node.Config{
+	stackConf := &node.Config{
 		DataDir:         MustMakeChainDataDir(ctx),
 		PrivateKey:      MakeNodeKey(ctx),
 		Name:            name,
@@ -515,8 +699,51 @@ func mustMakeStackConf(ctx *cli.Context, name string, config *core.SufficientCha
 		WSModules:       MakeRPCModules(ctx.GlobalString(aliasableName(WSApiFlag.Name, ctx))),
 	}
 
+	accman := MakeAccountManager(ctx)
+	passwords := MakePasswordList(ctx)
+
+	accounts := strings.Split(ctx.GlobalString(aliasableName(UnlockedAccountFlag.Name, ctx)), ",")
+	for i, account := range accounts {
+		if trimmed := strings.TrimSpace(account); trimmed != "" {
+			unlockAccount(ctx, accman, trimmed, i, passwords)
+		}
+	}
+
+	ethConf := &eth.Config{
+		ChainConfig:             config.ChainConfig,
+		FastSync:                ctx.GlobalBool(aliasableName(FastSyncFlag.Name, ctx)),
+		BlockChainVersion:       ctx.GlobalInt(aliasableName(BlockchainVersionFlag.Name, ctx)),
+		DatabaseCache:           ctx.GlobalInt(aliasableName(CacheFlag.Name, ctx)),
+		DatabaseHandles:         MakeDatabaseHandles(),
+		NetworkId:               ctx.GlobalInt(aliasableName(NetworkIdFlag.Name, ctx)),
+		AccountManager:          accman,
+		Etherbase:               MakeEtherbase(accman, ctx),
+		MinerThreads:            ctx.GlobalInt(aliasableName(MinerThreadsFlag.Name, ctx)),
+		NatSpec:                 ctx.GlobalBool(aliasableName(NatspecEnabledFlag.Name, ctx)),
+		DocRoot:                 ctx.GlobalString(aliasableName(DocRootFlag.Name, ctx)),
+		GasPrice:                new(big.Int),
+		GpoMinGasPrice:          new(big.Int),
+		GpoMaxGasPrice:          new(big.Int),
+		GpoFullBlockRatio:       ctx.GlobalInt(aliasableName(GpoFullBlockRatioFlag.Name, ctx)),
+		GpobaseStepDown:         ctx.GlobalInt(aliasableName(GpobaseStepDownFlag.Name, ctx)),
+		GpobaseStepUp:           ctx.GlobalInt(aliasableName(GpobaseStepUpFlag.Name, ctx)),
+		GpobaseCorrectionFactor: ctx.GlobalInt(aliasableName(GpobaseCorrectionFactorFlag.Name, ctx)),
+		SolcPath:                ctx.GlobalString(aliasableName(SolcPathFlag.Name, ctx)),
+		AutoDAG:                 ctx.GlobalBool(aliasableName(AutoDAGFlag.Name, ctx)) || ctx.GlobalBool(aliasableName(MiningEnabledFlag.Name, ctx)),
+	}
+
+	if _, ok := ethConf.GasPrice.SetString(ctx.GlobalString(aliasableName(GasPriceFlag.Name, ctx)), 0); !ok {
+		log.Fatalf("malformed %s flag value %q", aliasableName(GasPriceFlag.Name, ctx), ctx.GlobalString(aliasableName(GasPriceFlag.Name, ctx)))
+	}
+	if _, ok := ethConf.GpoMinGasPrice.SetString(ctx.GlobalString(aliasableName(GpoMinGasPriceFlag.Name, ctx)), 0); !ok {
+		log.Fatalf("malformed %s flag value %q", aliasableName(GpoMinGasPriceFlag.Name, ctx), ctx.GlobalString(aliasableName(GpoMinGasPriceFlag.Name, ctx)))
+	}
+	if _, ok := ethConf.GpoMaxGasPrice.SetString(ctx.GlobalString(aliasableName(GpoMaxGasPriceFlag.Name, ctx)), 0); !ok {
+		log.Fatalf("malformed %s flag value %q", aliasableName(GpoMaxGasPriceFlag.Name, ctx), ctx.GlobalString(aliasableName(GpoMaxGasPriceFlag.Name, ctx)))
+	}
+
 	// Configure the Whisper service
-	shhEnable = ctx.GlobalBool(aliasableName(WhisperEnabledFlag.Name, ctx))
+	shhEnable := ctx.GlobalBool(aliasableName(WhisperEnabledFlag.Name, ctx))
 
 	if ctx.GlobalIsSet(aliasableName(UseChainConfigFlag.Name, ctx)) {
 		ethConf.Genesis = config.Genesis // from parsed JSON file
@@ -552,63 +779,44 @@ func mustMakeStackConf(ctx *cli.Context, name string, config *core.SufficientCha
 		}
 		ethConf.PowTest = true
 	}
-
-	return stackConf, shhEnable
-}
-
-func mustMakeEthConf(ctx *cli.Context, sconf *core.SufficientChainConfig) *eth.Config {
-
-	accman := MakeAccountManager(ctx)
-	passwords := MakePasswordList(ctx)
-
-	accounts := strings.Split(ctx.GlobalString(aliasableName(UnlockedAccountFlag.Name, ctx)), ",")
-	for i, account := range accounts {
-		if trimmed := strings.TrimSpace(account); trimmed != "" {
-			unlockAccount(ctx, accman, trimmed, i, passwords)
+	// Assemble and return the protocol stack
+	stack, err := node.New(stackConf)
+	if err != nil {
+		glog.Fatalf("%v: failed to create the protocol stack: ", ErrStackFail, err)
+	}
+	if err := stack.Register(func(ctx *node.ServiceContext) (node.Service, error) {
+		return eth.New(ctx, ethConf)
+	}); err != nil {
+		glog.Fatalf("%v: failed to register the Ethereum service: ", ErrStackFail, err)
+	}
+	if shhEnable {
+		if err := stack.Register(func(*node.ServiceContext) (node.Service, error) { return whisper.New(), nil }); err != nil {
+			glog.Fatalf("%v: failed to register the Whisper service: ", ErrStackFail, err)
 		}
 	}
 
-	ethConf := &eth.Config{
-		ChainConfig:             sconf.ChainConfig,
-		FastSync:                ctx.GlobalBool(aliasableName(FastSyncFlag.Name, ctx)),
-		BlockChainVersion:       ctx.GlobalInt(aliasableName(BlockchainVersionFlag.Name, ctx)),
-		DatabaseCache:           ctx.GlobalInt(aliasableName(CacheFlag.Name, ctx)),
-		DatabaseHandles:         MakeDatabaseHandles(),
-		NetworkId:               ctx.GlobalInt(aliasableName(NetworkIdFlag.Name, ctx)),
-		AccountManager:          accman,
-		Etherbase:               MakeEtherbase(accman, ctx),
-		MinerThreads:            ctx.GlobalInt(aliasableName(MinerThreadsFlag.Name, ctx)),
-		NatSpec:                 ctx.GlobalBool(aliasableName(NatspecEnabledFlag.Name, ctx)),
-		DocRoot:                 ctx.GlobalString(aliasableName(DocRootFlag.Name, ctx)),
-		GasPrice:                new(big.Int),
-		GpoMinGasPrice:          new(big.Int),
-		GpoMaxGasPrice:          new(big.Int),
-		GpoFullBlockRatio:       ctx.GlobalInt(aliasableName(GpoFullBlockRatioFlag.Name, ctx)),
-		GpobaseStepDown:         ctx.GlobalInt(aliasableName(GpobaseStepDownFlag.Name, ctx)),
-		GpobaseStepUp:           ctx.GlobalInt(aliasableName(GpobaseStepUpFlag.Name, ctx)),
-		GpobaseCorrectionFactor: ctx.GlobalInt(aliasableName(GpobaseCorrectionFactorFlag.Name, ctx)),
-		SolcPath:                ctx.GlobalString(aliasableName(SolcPathFlag.Name, ctx)),
-		AutoDAG:                 ctx.GlobalBool(aliasableName(AutoDAGFlag.Name, ctx)) || ctx.GlobalBool(aliasableName(MiningEnabledFlag.Name, ctx)),
+	if ctx.GlobalBool(Unused1.Name) {
+		glog.V(logger.Info).Infoln(fmt.Sprintf("Geth started with --%s flag, which is unused by Geth Classic and can be omitted", Unused1.Name))
 	}
 
-	if _, ok := ethConf.GasPrice.SetString(ctx.GlobalString(aliasableName(GasPriceFlag.Name, ctx)), 0); !ok {
-		log.Fatalf("malformed %s flag value %q", aliasableName(GasPriceFlag.Name, ctx), ctx.GlobalString(aliasableName(GasPriceFlag.Name, ctx)))
-	}
-	if _, ok := ethConf.GpoMinGasPrice.SetString(ctx.GlobalString(aliasableName(GpoMinGasPriceFlag.Name, ctx)), 0); !ok {
-		log.Fatalf("malformed %s flag value %q", aliasableName(GpoMinGasPriceFlag.Name, ctx), ctx.GlobalString(aliasableName(GpoMinGasPriceFlag.Name, ctx)))
-	}
-	if _, ok := ethConf.GpoMaxGasPrice.SetString(ctx.GlobalString(aliasableName(GpoMaxGasPriceFlag.Name, ctx)), 0); !ok {
-		log.Fatalf("malformed %s flag value %q", aliasableName(GpoMaxGasPriceFlag.Name, ctx), ctx.GlobalString(aliasableName(GpoMaxGasPriceFlag.Name, ctx)))
-	}
-
-	return ethConf
+	return stack
 }
 
-// mustMakeSufficientChainConfig makes a sufficent chain configuration (id, chainconfig, nodes,...) from
+// MustMakeChainConfig reads the chain configuration from the database in ctx.Datadir.
+func MustMakeChainConfig(ctx *cli.Context) *core.ChainConfig {
+
+	db := MakeChainDatabase(ctx)
+	defer db.Close()
+	glog.V(logger.Info).Info(fmt.Sprintf("Did read chain configuration from database: \x1b[32m%s\x1b[39m", MustMakeChainDataDir(ctx)+"/chaindata"))
+
+	return MustMakeChainConfigFromDb(ctx, db)
+}
+
+// mustMakeSufficientConfiguration makes a sufficent chain configuration (id, chainconfig, nodes,...) from
 // either JSON file path, DB, or fails hard.
 // Forces users to provide a full and complete config file if any is specified.
 // Delegates flags to determine which source to use for configuration setup.
-func mustMakeSufficientChainConfig(ctx *cli.Context) *core.SufficientChainConfig {
+func mustMakeSufficientConfiguration(ctx *cli.Context) *core.SufficientChainConfig {
 
 	config := &core.SufficientChainConfig{}
 
@@ -623,7 +831,7 @@ func mustMakeSufficientChainConfig(ctx *cli.Context) *core.SufficientChainConfig
 
 		// Check for flagged bootnodes.
 		if ctx.GlobalIsSet(aliasableName(BootnodesFlag.Name, ctx)) {
-			glog.Fatal("Conflicting --chain-config and --bootnodes flags. Please use either but not both.")
+			panic("Conflicting --chain-config and --bootnodes flags. Please use either but not both.")
 		}
 
 		currentChainID = config.ID // Set global var.
@@ -634,14 +842,8 @@ func mustMakeSufficientChainConfig(ctx *cli.Context) *core.SufficientChainConfig
 	// Initialise chain configuration before handling migrations or setting up node.
 	config.ID = getChainConfigIDFromContext(ctx)
 	config.Name = getChainConfigNameFromContext(ctx)
-	config.ChainConfig = MustMakeChainConfigFromDefaults(ctx).SortForks()
+	config.ChainConfig = MustMakeChainConfig(ctx).SortForks()
 	config.ParsedBootstrap = MakeBootstrapNodesFromContext(ctx)
-
-	if isTestMode(ctx) {
-		config.Genesis = core.TestNetGenesis
-	} else {
-		config.Genesis = core.DefaultGenesis
-	}
 
 	return config
 }
@@ -649,22 +851,22 @@ func mustMakeSufficientChainConfig(ctx *cli.Context) *core.SufficientChainConfig
 func logChainConfiguration(ctx *cli.Context, config *core.SufficientChainConfig) {
 
 	if ctx.GlobalIsSet(aliasableName(UseChainConfigFlag.Name, ctx)) {
-		glog.V(logger.Info).Infof("Using custom chain configuration file: \x1b[32m%s\x1b[39m",
-			filepath.Clean(ctx.GlobalString(aliasableName(UseChainConfigFlag.Name, ctx))))
+		glog.V(logger.Info).Info(fmt.Sprintf("Using custom chain configuration file: \x1b[32m%s\x1b[39m",
+			filepath.Clean(ctx.GlobalString(aliasableName(UseChainConfigFlag.Name, ctx)))))
 	}
 
 	glog.V(logger.Info).Info(glog.Separator("-"))
 
-	glog.V(logger.Info).Infof("Starting Geth Classic \x1b[32m%s\x1b[39m", ctx.App.Version)
+	glog.V(logger.Info).Info(fmt.Sprintf("Starting Geth Classic \x1b[32m%s\x1b[39m", ctx.App.Version))
 	glog.V(logger.Info).Infof("Geth is configured to use blockchain: \x1b[32m%v (ETC)\x1b[39m", config.Name)
-	glog.V(logger.Info).Infof("Using chain database at: \x1b[32m%s\x1b[39m", MustMakeChainDataDir(ctx)+"/chaindata")
+	glog.V(logger.Info).Infof("Chain subdir: \x1b[32m%v\x1b[39m", config.ID)
 
-	glog.V(logger.Info).Infof("%v blockchain upgrades associated with this configuration:", len(config.ChainConfig.Forks))
+	glog.V(logger.Info).Info(fmt.Sprintf("%v blockchain upgrades associated with this configuration:", len(config.ChainConfig.Forks)))
 
 	for i := range config.ChainConfig.Forks {
-		glog.V(logger.Info).Infof(" %7v %v", config.ChainConfig.Forks[i].Block, config.ChainConfig.Forks[i].Name)
+		glog.V(logger.Info).Info(fmt.Sprintf(" %7v %v", config.ChainConfig.Forks[i].Block, config.ChainConfig.Forks[i].Name))
 		if !config.ChainConfig.Forks[i].RequiredHash.IsEmpty() {
-			glog.V(logger.Info).Infof("         with block %v", config.ChainConfig.Forks[i].RequiredHash.Hex())
+			glog.V(logger.Info).Info(fmt.Sprintf("         with block %v", config.ChainConfig.Forks[i].RequiredHash.Hex()))
 		}
 		for _, feat := range config.ChainConfig.Forks[i].Features {
 			glog.V(logger.Debug).Infof("    id: %v", feat.ID)
@@ -677,12 +879,18 @@ func logChainConfiguration(ctx *cli.Context, config *core.SufficientChainConfig)
 	glog.V(logger.Info).Info(glog.Separator("-"))
 }
 
-// MustMakeChainConfigFromDefaults reads the chain configuration from hardcode.
-func MustMakeChainConfigFromDefaults(ctx *cli.Context) *core.ChainConfig {
+// MustMakeChainConfigFromDb reads the chain configuration from the given database.
+func MustMakeChainConfigFromDb(ctx *cli.Context, db ethdb.Database) *core.ChainConfig {
+
 	c := core.DefaultConfig
+	configName := "mainnet"
 	if isTestMode(ctx) {
 		c = core.TestConfig
+		configName = "morden testnet"
 	}
+
+	glog.V(logger.Info).Info(fmt.Sprintf("Loading \x1b[36m%v\x1b[39m configuration from database...", configName))
+
 	return c
 }
 
@@ -704,14 +912,15 @@ func MakeChainDatabase(ctx *cli.Context) ethdb.Database {
 // MakeChain creates a chain manager from set command line flags.
 func MakeChain(ctx *cli.Context) (chain *core.BlockChain, chainDb ethdb.Database) {
 	var err error
-	sconf := mustMakeSufficientChainConfig(ctx)
 	chainDb = MakeChainDatabase(ctx)
+
+	chainConfig := MustMakeChainConfigFromDb(ctx, chainDb)
 
 	pow := pow.PoW(core.FakePow{})
 	if !ctx.GlobalBool(aliasableName(FakePoWFlag.Name, ctx)) {
 		pow = ethash.New()
 	}
-	chain, err = core.NewBlockChain(chainDb, sconf.ChainConfig, pow, new(event.TypeMux))
+	chain, err = core.NewBlockChain(chainDb, chainConfig, pow, new(event.TypeMux))
 	if err != nil {
 		glog.Fatal("Could not start chainmanager: ", err)
 	}
